pyproject.toml
[project]
name = "agentlab"
version = "0.1.0"
description = "Agentlab project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
